+++ core/server/api_container/server/startosis_engine/startosis_const.go
@@ -6,5 +6,5 @@ const (
 	MainInputArgName = "args"
 	EmptyInputArgs   = "{}" // empty JSON
 
-	ModuleIdPlaceholderForStandaloneScripts = "DEFAULT_MODULE_ID_FOR_SCRIPT"
+	PackageIdPlaceholderForStandaloneScript = "DEFAULT_PACKAGE_ID_FOR_SCRIPT"
 )