+++ internal_testsuites/golang/testsuite/startosis_remove_service_test/startosis_remove_service_test.go
@@ -54,7 +54,7 @@ func TestStartosis(t *testing.T) {
 	logrus.Infof("Executing Startosis script...")
 	logrus.Debugf("Startosis script content: \n%v", startosisScript)
 
-	outputStream, _, err := enclaveCtx.ExecuteKurtosisScript(ctx, startosisScript, defaultDryRun)
+	outputStream, _, err := enclaveCtx.RunStarlarkScript(ctx, startosisScript, defaultDryRun)
 	require.NoError(t, err, "Unexpected error executing startosis script")
 	scriptOutput, _, interpretationError, validationErrors, executionError := test_helpers.ReadStreamContentUntilClosed(outputStream)
 
@@ -79,7 +79,7 @@ Service example-datastore-server-1 deployed successfully.
 	logrus.Infof("All services added via the module work as expected")
 
 	// we run the remove script and see if things still work
-	outputStream, _, err = enclaveCtx.ExecuteKurtosisScript(ctx, removeScript, defaultDryRun)
+	outputStream, _, err = enclaveCtx.RunStarlarkScript(ctx, removeScript, defaultDryRun)
 	require.NoError(t, err, "Unexpected error executing remove script")
 	_, _, interpretationError, validationErrors, executionError = test_helpers.ReadStreamContentUntilClosed(outputStream)
 	require.Nil(t, interpretationError, "Unexpected interpretation error")