+++ internal_testsuites/typescript/src/testsuite/startosis_module_test/shared_constants.ts
@@ -1,4 +1,0 @@
-export const IS_PARTITIONING_ENABLED = false;
-export const EMPTY_EXECUTE_PARAMS = "{}"
-export const DEFAULT_DRY_RUN = false;
-export const JEST_TIMEOUT_MS = 180000